tools/requirements.txt
pyserial
esptool
fastapi
uvicorn
jinja2
python-multipart
websockets
